[DSPDC-1699] Rename HCA load pipeline and add stub of datafile load branch (#110)
Why

We need to implement the "load data files" branch of the HCA load pipeline in Dagster.
This is the first of multiple PRs to implement this functionality.

This PR

Renames the stage_data pipeline to load_hca and updates references
Creates a load_data_files solid that is the stub of the branch to be implemented
