+++ orchestration/dagster_orchestration/hca_orchestration/pipelines/__init__.py
@@ -1,10 +1,10 @@
 from hca_orchestration.pipelines.cut_snapshot import cut_snapshot
-from hca_orchestration.pipelines.stage_data import stage_data
+from hca_orchestration.pipelines.load_hca import load_hca
 from hca_orchestration.pipelines.validate_egress import validate_egress
 
 
 __all__ = [
     'cut_snapshot',
-    'stage_data',
+    'load_hca',
     'validate_egress',
 ]
