+++ orchestration/dagster_orchestration/hca_orchestration/pipelines/load_hca.py
@@ -1,7 +1,8 @@
 from dagster import ModeDefinition, pipeline
 
 from hca_orchestration.config import preconfigure_resource_for_mode
-from hca_orchestration.solids.stage_data import clear_staging_dir, pre_process_metadata, create_staging_dataset
+from hca_orchestration.solids.load_hca.stage_data import clear_staging_dir, pre_process_metadata, create_staging_dataset
+from hca_orchestration.solids.load_hca.load_data_files import import_data_files
 from hca_orchestration.resources import dataflow_beam_runner, local_beam_runner, google_storage_client, \
     jade_data_repo_client, test_beam_runner, local_storage_client, noop_data_repo_client, bigquery_client, \
     noop_bigquery_client, load_tag
@@ -61,5 +62,6 @@ test_mode = ModeDefinition(
 @pipeline(
     mode_defs=[prod_mode, dev_mode, local_mode, test_mode]
 )
-def stage_data() -> None:
-    create_staging_dataset(pre_process_metadata(clear_staging_dir()))
+def load_hca() -> None:
+    staging_dataset = create_staging_dataset(pre_process_metadata(clear_staging_dir()))
+    import_data_files(staging_dataset)
