+++ orchestration/dagster_orchestration/hca_orchestration/solids/load_hca/stage_data.py
@@ -4,6 +4,8 @@ from dagster import solid, InputDefinition, Nothing, String, Int, OutputDefiniti
 from dagster.core.execution.context.compute import AbstractComputeExecutionContext
 from google.cloud.bigquery import Dataset
 
+from hca_orchestration.support.typing import HcaStagingDatasetName
+
 
 @solid(
     required_resource_keys={"storage_client", "staging_bucket_config"},
@@ -60,9 +62,8 @@ def pre_process_metadata(context: AbstractComputeExecutionContext) -> Nothing:
         "staging_table_expiration_ms": Int
     },
     input_defs=[InputDefinition("start", Nothing)],
-    output_defs=[OutputDefinition(name="staging_dataset_name", dagster_type=String)]
 )
-def create_staging_dataset(context: AbstractComputeExecutionContext) -> str:
+def create_staging_dataset(context: AbstractComputeExecutionContext) -> HcaStagingDatasetName:
     """
     Creates a staging dataset that will house records for update/insertion into the
     final TDR dataset
@@ -82,4 +83,4 @@ def create_staging_dataset(context: AbstractComputeExecutionContext) -> str:
 
     context.log.info(f"Created staging dataset {dataset_name}")
 
-    return dataset_name
+    return HcaStagingDatasetName(dataset_name)
